8/1 Update page configuration
