+++ client/src/components/shared/goodsNavbar/goodsNavbar.js
@@ -1,20 +1,40 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
+import { useSelector } from 'react-redux';
+
 import { Nav } from './styles';
 import { Logo } from '../../../assets/commonStyle/styles';
 import { faShoppingCart } from '@fortawesome/free-solid-svg-icons';
 import { faUser } from '@fortawesome/free-regular-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-const GoodsNavbar = (page) => {
+const GoodsNavbar = (nav) => {
+  const page = 'goods';
+  const { isAuthenticated } = useSelector((state) => state.auth);
   return (
     <Nav className="navBar">
       <div>
-        <Logo />
-        <a href="/" className="navLink navItem" id="memories">
+        {Logo(page)}
+        {nav === 'donation' ? (
+          <Link
+            to="/donation"
+            className="navLink navItem nav_active"
+            id="donation"
+          >
+            Donation
+            <p></p>
+          </Link>
+        ) : (
+          <Link to="/donation" className="navLink navItem" id="donation">
+            Donation
+            <p></p>
+          </Link>
+        )}
+        <Link to="/" className="navLink navItem" id="memories">
           Memories
           <p></p>
-        </a>
-        <a href="/shopping" className="navItem">
+        </Link>
+        <Link to="/shopping" className="navItem">
           <FontAwesomeIcon
             className="nav-item"
             icon={faShoppingCart}
@@ -27,20 +47,24 @@ const GoodsNavbar = (page) => {
               marginTop: '45px',
             }}
           />
-        </a>
-        <a className="navItem">
-          <FontAwesomeIcon
-            className="nav-item"
-            icon={faUser}
-            style={{
-              fontSize: '25px',
-              position: 'absolute',
-              right: '70px',
-              color: '#A4A4A4',
-              marginTop: '45px',
-            }}
-          />
-        </a>
+        </Link>
+        {isAuthenticated ? (
+          <Link className="navItem">
+            <FontAwesomeIcon
+              className="nav-item"
+              icon={faUser}
+              style={{
+                fontSize: '25px',
+                position: 'absolute',
+                right: '70px',
+                color: '#A4A4A4',
+                marginTop: '45px',
+              }}
+            />
+          </Link>
+        ) : (
+          <div></div>
+        )}
       </div>
     </Nav>
   );
