+++ client/src/pages/Shopping/styles.js
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 
 export const Page = styled.div`
   width: 100%;
-  height: 200vh;
+  height: 100vh;
 `;
 
 export const BuyArea = styled.section`
