+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "no_name_project",
+  "name": "memories_project",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
